Fixed the issue #80 (incorrect binary operations on vector-valued results).
The problem was that the vector of jackknife bins was recreated
(emptied+resized), and the vector elements (that are vectors
themselves) were default-initialized, that is, becoming vectors of
size 0. The fix is to treat any empty (presumably,
default-initialized) vectors as zeros of the field.

Also, some unused code in numeric/vector-functions.hpp is
commented-out by this commit.