Making custom accumulator value types easier to use.
To use a custom value type, take `accumulators/test/custom_type.hpp`
as a template to follow. The idea is to define
`ALPS_ACCUMULATOR_USER_TYPES` before including `alps/config.hpp`, then
make sure that your value type follows the concept as described in the
`custom_type.hpp`. The methods marked as `@REQUIRED`,
`@REQURED_SEQUENCE` and `@REQUIRED_NONCONTINUOUS` are part of the
concept, the rest are implementation details.

This commit is the essence of resolution of issue #157: it is now
possible to define a custom type containing Eigen matrix in a similar way.