Added proof-of-concept `apply`
This is a test and a proof-of-concept implementation of old `apply`
free function that used to access parameters metadata. As default parameter
value is not accessible as a distinct value once the parameter is
defined, the test currently fails. As per discussion in issue #359,
this functionality is withheld for now.

This commit is for archiving purposes only and will be reverted.