New CMake variable setting `ALPS_CXX_STD`
Setting '-DALPS_CXX_STD=c++03` or '-DALPS_CXX_STD=c++11` chooses C++03
or C++11 standard, respectively. This choice is enforced on the client
code.

Together with enforcement of C++ compiler ID and version, this should
take care of C++ standard compatibility (unless the user tries to
override the standard in the client code).

This essentially addresses the issue #301.