Add interface for adding a bundle of data.
NB this is *not* the same thing as adding data with a weight; rather,
it is close to adding a bundle, which is assumed to have the count
associated.  The interface is thus protected, one might down the line
add a function like this:

    struct computed_bundle<T> { computed<T> src; size_t weight; };
    ...
    x_acc &operator<<(const computed_bundle<T> &src);

This is useful in (a) fixing the autocorr hierarchy; (b) re-use of
lower-lying results in `batch_result`.