Properly rescale (co-)variances before applying transform.
Also adds a test.

This issue is the side product of the fact that the variance of
a set of weighted measurements and the "ensemble" variance differ
by a factor, so we have to correct for that factor in the computation
of the variance, which means we have to undo that before we
transform that variance.  The clean solution would be to have a
"batching decorator" that makes this transformation.

Fixes: #506