Allow client CMake to search for Boost; remember Boost version
This should close #552.

However, this is not a perfect solution: it is still possible for
CMake `find_package(Boost ${ALPSCore_BOOST_VERSION} EXACT)` to find
Boost libraries in the wrong location suggested by `BOOST_ROOT`
environment or CMAke variable. It looks like a CMake (or, rather,
`FindBoost.cmake`) bug to me.