Cmdline overrides params restored from HDF5
If there is a single HDF5 file as a command line argument, the
parameters object is restored from that file. After that, the remaining `key=value`
pairs from the command line are parsed according to the types of the
values associated with the `key`s in the restored parameters
object. If parsing fails, exception is thrown.

This should close #436.